requirements.txt
google-generativeai
python-dotenv
langchain-community 
langgraph
langchain-google-genai
langchain-chroma
chardet
requests
jq
tavily-python
fastapi
loguru
jinja2